+++ src/pages/Commits.tsx
@@ -13,7 +13,7 @@ import {
 import { GET_MAINLINE_COMMITS } from "gql/queries";
 import { usePageTitle, useNetworkStatus } from "hooks";
 import { PageDoesNotExist } from "pages/404";
-import { CommitGraphWrapper } from "pages/commits/commitGraph/CommitsWrapper";
+import { CommitGraphWrapper } from "pages/commits/CommitsWrapper";
 import { ProjectFilterOptions } from "types/commits";
 import { ProjectSelect } from "./commits/ProjectSelect";
